Ensure helper classes are injected properly
Nested classes that don't have accessible parents can cause problems, so ensure that they are also injected with muzzle.

Fix various failures.

(Also update muzzle task definition in effort to improve up to date check.)
