+++ dd-java-agent/instrumentation/hibernate/core-3.3/src/main/java/datadog/trace/instrumentation/hibernate/core/v3_3/AbstractHibernateInstrumentation.java
@@ -3,8 +3,6 @@ package datadog.trace.instrumentation.hibernate.core.v3_3;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.instrumentation.hibernate.HibernateMatchers;
 import net.bytebuddy.matcher.ElementMatcher;
-import org.hibernate.classic.Validatable;
-import org.hibernate.transaction.JBossTransactionManagerLookup;
 
 public abstract class AbstractHibernateInstrumentation extends Instrumenter.Tracing {
 
@@ -24,23 +22,6 @@ public abstract class AbstractHibernateInstrumentation extends Instrumenter.Trac
       "datadog.trace.instrumentation.hibernate.SessionMethodUtils",
       "datadog.trace.instrumentation.hibernate.SessionState",
       "datadog.trace.instrumentation.hibernate.HibernateDecorator",
-      packageName + ".AbstractHibernateInstrumentation$V3Advice",
     };
-  }
-
-  public abstract static class V3Advice {
-
-    /**
-     * Some cases of instrumentation will match more broadly than others, so this unused method
-     * allows all instrumentation to uniformly match versions of Hibernate between 3.3 and 4.
-     */
-    public static void muzzleCheck(
-        // Not in 4.0
-        final Validatable validatable,
-        // Not before 3.3.0.GA
-        final JBossTransactionManagerLookup lookup) {
-      validatable.validate();
-      lookup.getUserTransactionName();
-    }
   }
 }
