+++ dd-java-agent/instrumentation/hibernate/core-4.0/src/main/java/datadog/trace/instrumentation/hibernate/core/v4_0/AbstractHibernateInstrumentation.java
@@ -3,7 +3,6 @@ package datadog.trace.instrumentation.hibernate.core.v4_0;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.instrumentation.hibernate.HibernateMatchers;
 import net.bytebuddy.matcher.ElementMatcher;
-import org.hibernate.SharedSessionContract;
 
 public abstract class AbstractHibernateInstrumentation extends Instrumenter.Tracing {
 
@@ -23,18 +22,6 @@ public abstract class AbstractHibernateInstrumentation extends Instrumenter.Trac
       "datadog.trace.instrumentation.hibernate.SessionMethodUtils",
       "datadog.trace.instrumentation.hibernate.SessionState",
       "datadog.trace.instrumentation.hibernate.HibernateDecorator",
-      packageName + ".AbstractHibernateInstrumentation$V4Advice",
     };
-  }
-
-  public abstract static class V4Advice {
-
-    /**
-     * Some cases of instrumentation will match more broadly than others, so this unused method
-     * allows all instrumentation to uniformly match versions of Hibernate starting at 4.0.
-     */
-    public static void muzzleCheck(final SharedSessionContract contract) {
-      contract.createCriteria("");
-    }
   }
 }
