+++ dd-java-agent/instrumentation/hibernate/core-4.0/src/main/java/datadog/trace/instrumentation/hibernate/core/v4_0/CriteriaInstrumentation.java
@@ -20,6 +20,7 @@ import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.implementation.bytecode.assign.Assigner;
 import net.bytebuddy.matcher.ElementMatcher;
 import org.hibernate.Criteria;
+import org.hibernate.SharedSessionContract;
 
 @AutoService(Instrumenter.class)
 public class CriteriaInstrumentation extends AbstractHibernateInstrumentation {
@@ -41,7 +42,7 @@ public class CriteriaInstrumentation extends AbstractHibernateInstrumentation {
         CriteriaInstrumentation.class.getName() + "$CriteriaMethodAdvice");
   }
 
-  public static class CriteriaMethodAdvice extends V4Advice {
+  public static class CriteriaMethodAdvice {
 
     @Advice.OnMethodEnter(suppress = Throwable.class)
     public static SessionState startMethod(
@@ -63,6 +64,14 @@ public class CriteriaInstrumentation extends AbstractHibernateInstrumentation {
         @Advice.Return(typing = Assigner.Typing.DYNAMIC) final Object entity) {
 
       SessionMethodUtils.closeScope(state, throwable, entity, true);
+    }
+
+    /**
+     * Some cases of instrumentation will match more broadly than others, so this unused method
+     * allows all instrumentation to uniformly match versions of Hibernate starting at 4.0.
+     */
+    public static void muzzleCheck(final SharedSessionContract contract) {
+      contract.createCriteria("");
     }
   }
 }
