+++ dd-java-agent/instrumentation/hibernate/core-4.0/src/main/java/datadog/trace/instrumentation/hibernate/core/v4_0/QueryInstrumentation.java
@@ -22,6 +22,7 @@ import net.bytebuddy.implementation.bytecode.assign.Assigner;
 import net.bytebuddy.matcher.ElementMatcher;
 import org.hibernate.Query;
 import org.hibernate.SQLQuery;
+import org.hibernate.SharedSessionContract;
 
 @AutoService(Instrumenter.class)
 public class QueryInstrumentation extends AbstractHibernateInstrumentation {
@@ -43,7 +44,7 @@ public class QueryInstrumentation extends AbstractHibernateInstrumentation {
         QueryInstrumentation.class.getName() + "$QueryMethodAdvice");
   }
 
-  public static class QueryMethodAdvice extends V4Advice {
+  public static class QueryMethodAdvice {
 
     @Advice.OnMethodEnter(suppress = Throwable.class)
     public static SessionState startMethod(
@@ -78,6 +79,14 @@ public class QueryInstrumentation extends AbstractHibernateInstrumentation {
       }
 
       SessionMethodUtils.closeScope(state, throwable, entity, true);
+    }
+
+    /**
+     * Some cases of instrumentation will match more broadly than others, so this unused method
+     * allows all instrumentation to uniformly match versions of Hibernate starting at 4.0.
+     */
+    public static void muzzleCheck(final SharedSessionContract contract) {
+      contract.createCriteria("");
     }
   }
 }
