+++ dd-java-agent/instrumentation/hystrix-1.4/src/main/java/datadog/trace/instrumentation/hystrix/HystrixInstrumentation.java
@@ -4,14 +4,11 @@ import static datadog.trace.agent.tooling.ClassLoaderMatcher.hasClassesNamed;
 import static datadog.trace.agent.tooling.bytebuddy.matcher.DDElementMatchers.extendsClass;
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.namedOneOf;
-import static datadog.trace.instrumentation.hystrix.HystrixDecorator.DECORATE;
 import static net.bytebuddy.matcher.ElementMatchers.returns;
 
 import com.google.auto.service.AutoService;
 import com.netflix.hystrix.HystrixInvokableInfo;
 import datadog.trace.agent.tooling.Instrumenter;
-import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
-import datadog.trace.instrumentation.rxjava.TracedOnSubscribe;
 import java.util.HashMap;
 import java.util.Map;
 import net.bytebuddy.asm.Advice;
@@ -22,8 +19,6 @@ import rx.Observable;
 
 @AutoService(Instrumenter.class)
 public class HystrixInstrumentation extends Instrumenter.Tracing {
-
-  private static final String OPERATION_NAME = "hystrix.cmd";
 
   public HystrixInstrumentation() {
     super("hystrix");
@@ -52,7 +47,7 @@ public class HystrixInstrumentation extends Instrumenter.Tracing {
       packageName + ".HystrixDecorator",
       packageName + ".HystrixDecorator$1",
       packageName + ".HystrixDecorator$ResourceNameCacheKey",
-      packageName + ".HystrixInstrumentation$HystrixOnSubscribe",
+      packageName + ".HystrixOnSubscribe",
     };
   }
 
@@ -89,28 +84,6 @@ public class HystrixInstrumentation extends Instrumenter.Tracing {
         @Advice.Thrown final Throwable throwable) {
 
       result = Observable.create(new HystrixOnSubscribe(result, command, "fallback"));
-    }
-  }
-
-  public static class HystrixOnSubscribe extends TracedOnSubscribe {
-    private final HystrixInvokableInfo<?> command;
-    private final String methodName;
-
-    public HystrixOnSubscribe(
-        final Observable originalObservable,
-        final HystrixInvokableInfo<?> command,
-        final String methodName) {
-      super(originalObservable, OPERATION_NAME, DECORATE);
-
-      this.command = command;
-      this.methodName = methodName;
-    }
-
-    @Override
-    protected void afterStart(final AgentSpan span) {
-      super.afterStart(span);
-
-      DECORATE.onCommand(span, command, methodName);
     }
   }
 }
