+++ dd-java-agent/instrumentation/jax-rs-client-2.0/connection-error-handling-resteasy/src/main/java/datadog/trace/instrumentation/connection_error/resteasy/ResteasyClientConnectionErrorInstrumentation.java
@@ -11,10 +11,7 @@ import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
 import datadog.trace.instrumentation.jaxrs.ClientTracingFilter;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
@@ -40,7 +37,7 @@ public final class ResteasyClientConnectionErrorInstrumentation extends Instrume
   @Override
   public String[] helperClassNames() {
     return new String[] {
-      getClass().getName() + "$WrappedFuture",
+      packageName + ".WrappedFuture",
     };
   }
 
@@ -85,65 +82,6 @@ public final class ResteasyClientConnectionErrorInstrumentation extends Instrume
         @Advice.Return(readOnly = false) Future<?> future) {
       if (!(future instanceof WrappedFuture)) {
         future = new WrappedFuture<>(future, context);
-      }
-    }
-  }
-
-  public static class WrappedFuture<T> implements Future<T> {
-
-    private final Future<T> wrapped;
-    private final ClientConfiguration context;
-
-    public WrappedFuture(final Future<T> wrapped, final ClientConfiguration context) {
-      this.wrapped = wrapped;
-      this.context = context;
-    }
-
-    @Override
-    public boolean cancel(final boolean mayInterruptIfRunning) {
-      return wrapped.cancel(mayInterruptIfRunning);
-    }
-
-    @Override
-    public boolean isCancelled() {
-      return wrapped.isCancelled();
-    }
-
-    @Override
-    public boolean isDone() {
-      return wrapped.isDone();
-    }
-
-    @Override
-    public T get() throws InterruptedException, ExecutionException {
-      try {
-        return wrapped.get();
-      } catch (final ExecutionException e) {
-        final Object prop = context.getProperty(ClientTracingFilter.SPAN_PROPERTY_NAME);
-        if (prop instanceof AgentSpan) {
-          final AgentSpan span = (AgentSpan) prop;
-          span.setError(true);
-          span.addThrowable(e.getCause());
-          span.finish();
-        }
-        throw e;
-      }
-    }
-
-    @Override
-    public T get(final long timeout, final TimeUnit unit)
-        throws InterruptedException, ExecutionException, TimeoutException {
-      try {
-        return wrapped.get(timeout, unit);
-      } catch (final ExecutionException e) {
-        final Object prop = context.getProperty(ClientTracingFilter.SPAN_PROPERTY_NAME);
-        if (prop instanceof AgentSpan) {
-          final AgentSpan span = (AgentSpan) prop;
-          span.setError(true);
-          span.addThrowable(e.getCause());
-          span.finish();
-        }
-        throw e;
       }
     }
   }
