+++ dd-java-agent/instrumentation/opentracing/api-0.32/src/main/java/datadog/trace/instrumentation/opentracing32/GlobalTracerInstrumentation.java
@@ -36,7 +36,7 @@ public class GlobalTracerInstrumentation extends Instrumenter.Tracing {
     return new String[] {
       packageName + ".OTTracer",
       packageName + ".OTTracer$OTSpanBuilder",
-      packageName + ".OTPropagation$TextMapInjectSetter",
+      packageName + ".OTTextMapInjectSetter",
       packageName + ".OTScopeManager",
       packageName + ".OTScopeManager$OTScope",
       packageName + ".OTScopeManager$OTTraceScope",
