--- dd-java-agent/instrumentation/opentracing/api-0.32/src/main/java/datadog/trace/instrumentation/opentracing32/OTPropagation.java
@@ -1,16 +1,0 @@
-package datadog.trace.instrumentation.opentracing32;
-
-import datadog.trace.bootstrap.instrumentation.api.AgentPropagation;
-import io.opentracing.propagation.TextMapInject;
-
-class OTPropagation {
-
-  static class TextMapInjectSetter implements AgentPropagation.Setter<TextMapInject> {
-    static final TextMapInjectSetter INSTANCE = new TextMapInjectSetter();
-
-    @Override
-    public void set(final TextMapInject carrier, final String key, final String value) {
-      carrier.put(key, value);
-    }
-  }
-}
