+++ dd-java-agent/instrumentation/opentracing/api-0.32/src/main/java/datadog/trace/instrumentation/opentracing32/OTTextMapInjectSetter.java
@@ -1,0 +1,13 @@
+package datadog.trace.instrumentation.opentracing32;
+
+import datadog.trace.bootstrap.instrumentation.api.AgentPropagation;
+import io.opentracing.propagation.TextMapInject;
+
+class OTTextMapInjectSetter implements AgentPropagation.Setter<TextMapInject> {
+  static final OTTextMapInjectSetter INSTANCE = new OTTextMapInjectSetter();
+
+  @Override
+  public void set(final TextMapInject carrier, final String key, final String value) {
+    carrier.put(key, value);
+  }
+}
